Backend/srtm & motionOfWaves/requirements.txt
# SRTM & Motion Detection API - Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
python-multipart==0.0.6

# DEM/GeoTIFF Processing
rasterio==1.3.9
numpy==1.26.3
scipy==1.12.0

# Image/Video Processing
opencv-python==4.9.0.80

# Visualization
matplotlib==3.8.2

# Validation
pydantic==2.5.3
